feat: move freezing validation
